AI_SERVICE/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
Pillow
numpy
python-consul2
